Feature/adjust deleted users

Closes #18

See merge request !34

+++ app/javascript/packs/users/users_table_item.jsx
@@ -10,6 +10,7 @@ class UsersTableItem extends React.Component {
   }
 
   renderCardSerial(value) {
+    const user = this.props.user;
     if (value.length > 0) {
       return value;
     }
@@ -19,7 +20,7 @@ class UsersTableItem extends React.Component {
       this.registerUser();
     };
 
-    if (this.props.isAdmin === 'true') {
+    if (this.props.isAdmin === 'true' && !user.deleted) {
       return (
         <a href="" onClick={onClick} className="btn btn-success">綁定</a>
       );
@@ -29,16 +30,26 @@ class UsersTableItem extends React.Component {
   }
 
   renderEditAndDeleteButton(links) {
+    const view = [];
+    const user = this.props.user;
     if (this.props.isAdmin === 'true') {
-      return ([
-        <a href={links.edit} className="btn btn-primary">編輯</a>,
-        <a
+      view.push(<a href={links.edit} className="btn btn-primary">編輯</a>);
+      if (user.deleted) {
+        view.push(<a
+          href={links.recover}
+          className="btn btn-success"
+          data-method="patch"
+          data-confirm="Are you sure?"
+        >復職</a>);
+      } else {
+        view.push(<a
           href={links.self}
           className="btn btn-danger"
           data-method="delete"
           data-confirm="Are you sure?"
-        >刪除</a>,
-      ]);
+        >離職</a>);
+      }
+      return view;
     }
 
     return undefined;
